+++ truss/cli/common.py
@@ -1,0 +1,2 @@
+
+POLL_INTERVAL_SEC = 2
