+++ truss/cli/train.py
@@ -5,7 +5,7 @@ import rich_click as click
 from InquirerPy import inquirer
 from rich.console import Console
 
-from truss.cli.training_metrics_watcher import MetricsWatcher
+from truss.cli.metrics.training_metrics_watcher import MetricsWatcher
 from truss.remote.baseten.remote import BasetenRemote
 
 ACTIVE_JOB_STATUSES = [
@@ -200,6 +200,5 @@ def view_training_job_metrics(
     view_training_job_metrics shows a list of metrics for a training job.
     """
     project_id, job_id = get_args_for_logs(console, remote_provider, project_id, job_id)
-    console.print(f"Showing metrics for job: {job_id}")
     metrics_display = MetricsWatcher(remote_provider.api, project_id, job_id, console)
     metrics_display.display_live_metrics()
